// studio/sanity.types.ts
export declare const internalGroqTypeReferenceTo: unique symbol
// ...
export type PageReference = {
  _ref: string
  _type: 'reference'
  _weak?: boolean
  [internalGroqTypeReferenceTo]?: 'page'
}
// ...
export type PostReference = {
  _ref: string
  _type: 'reference'
  _weak?: boolean
  [internalGroqTypeReferenceTo]?: 'post'
}

export type Link = {
  _type: 'link'
  linkType?: 'href' | 'page' | 'post'
  href?: string
  page?: PageReference
  post?: PostReference
  openInNewTab?: boolean
}